requirements.txt
beautifulsoup4
langchain
langchain-community
langchain-anthropic
langchain-chroma
langchain-google-genai
langchain-experimental
langchain-groq
langchain-openai
langfuse
matplotlib
neo4j
pymupdf
pypdf
python-docx
python-dotenv
streamlit
streamlit-agraph
streamlit-option-menu
transformers
typing-inspect
unstructured
